added feature to create news post, delete them, and show them to users

+++ avanzareApp/views.py
@@ -1,10 +1,11 @@
 from flask import Blueprint, render_template, request, flash, jsonify, redirect, url_for
 from flask_login import login_required, current_user
-from .models import Note, Menu, User, Order, Menu_order
+from .models import Note, Menu, User, Order, Menu_order, News
 from . import db
 import json
 
 views = Blueprint('views', __name__)
+# handling floating point errors
 
 # add orders variable to User class
 # create a class ofr orders and link by id
@@ -101,8 +102,9 @@ def user_home():
     error_orders = Order.query.filter(Order.user_id == current_user.id).filter(Order.total == 0).delete()
     db.session.commit()
     active_orders = Order.query.filter(Order.user_id == current_user.id).filter(Order.is_active == "Active").all()
+    news = db.session.query(News).order_by(News.id.desc()).all()
     db.session.commit()
-    return render_template("user_home.html", active_orders=active_orders, user=current_user)
+    return render_template("user_home.html", news=news, active_orders=active_orders, user=current_user)
 
 @views.route('/order', methods=['GET', 'POST'])
 @login_required
@@ -114,7 +116,8 @@ def order():
     db.session.commit()
     # create order
     # we need to pass new_item to access its order id
-    new_order = Order(user_id=current_user.id, total=0)
+    # MAYBE ITEMS=[] EMPTY LIST WILL FIX BUG WHERE IT SHOWS ITEMS WHEN YOU CLICK IN AND OUT OF ORDER
+    new_order = Order(user_id=current_user.id, taxes=0,total=0)
     db.session.add(new_order)
     db.session.commit()
 
@@ -130,6 +133,7 @@ def add_to_order(order_id):
     quantity = request.form.get('quantity')
     dish_accommodation = request.form.get('dish_accommodation')
     complete_order = request.form.get('complete_order')
+
     # just in case someone enters spaces but no digits
     # if quantity == none then can't strip usually when fresh order this occurs
     if quantity != None:
@@ -145,15 +149,36 @@ def add_to_order(order_id):
         item = Menu_order.query.filter(Menu_order.id == item_id2).first()
         new_order = Order.query.filter(Order.id == order_id).first()
         delete_quantity = int(request.form.get('delete_quantity'))
+        # if page is refreshed after deleting something then an errors occurs because item  becomes none
+        # and tmp_name = item.name[:] causes and issue
+        # so in that case refresh page
+        if item is None:
+            user = User.query.filter(User.email == 'headChef@gmail.com').first()
+            items = Menu.query.filter_by(user_id=user.id).order_by(Menu.name).all()
+            return render_template('order.html', items=items, new_order=new_order, user=current_user)
         # if you delete the item then you can't use it to delete other items
         tmp_name = item.name[:]
+        count_lst = []
+        for item_tmp in new_order.items:
+            count_lst.append(item_tmp.name)
+
+        count = count_lst.count(item.name)
+        if count < delete_quantity:
+            flash('Delete quantity greater than the amounts of items in order', category="error")
+            user = User.query.filter(User.email == 'headChef@gmail.com').first()
+            items = Menu.query.filter_by(user_id=user.id).order_by(Menu.name).all()
+            return render_template('order.html', items=items, new_order=new_order, user=current_user)
+
         for x in range(delete_quantity):
             for item_compare_delete in new_order.items:
                 if tmp_name == item_compare_delete.name:
+                    new_order.total = new_order.total - float(item_compare_delete.price)
                     db.session.delete(item_compare_delete)
                     db.session.commit()
                     break
-
+        # calculate taxes and add to total
+        new_order.taxes = new_order.total * 0.08
+        db.session.commit()
         flash('ITEMS DELETED!', category="success")
         user = User.query.filter(User.email == 'headChef@gmail.com').first()
         items = Menu.query.filter_by(user_id=user.id).order_by(Menu.name).all()
@@ -182,9 +207,12 @@ def add_to_order(order_id):
             item_copy.order_id = order_id
             # calculate order total
             new_order.total = new_order.total + float(item_copy.price)
+
             db.session.add(item_copy)
             db.session.commit()
 
+        new_order.taxes = new_order.total * 0.08
+        db.session.commit()
         flash('Item added.', category="success")
         user = User.query.filter(User.email == 'headChef@gmail.com').first()
         items = Menu.query.filter_by(user_id=user.id).order_by(Menu.name).all()
@@ -443,4 +471,30 @@ def order_again(id):
         return redirect(url_for('views.user_home'))
 
 
+@views.route('/create-news', methods=['GET', 'POST'])
+@login_required
+def create_news():
+    if request.method == "POST":
+        title = request.form.get('title')
+        text = request.form.get('text')
+        news_type = request.form.get('news_type')
+        news_post = News(title=title, text=text, news_type=news_type)
+        db.session.add(news_post)
+        db.session.commit()
+        flash('News post created!', category="success")
+    news = db.session.query(News).order_by(News.id.desc()).all()
+    return render_template("create_news.html", news=news, user=current_user)
 
+@views.route('/delete-news-post', methods=['POST', 'GET'])
+def delete_news_post():
+    news_post = json.loads(request.data)
+    news_postId = news_post['news_postId']
+    news_post = News.query.get(news_postId)
+
+    if news_post:
+        db.session.delete(news_post)
+        db.session.commit()
+        flash('NEWS POST DELETED!', category="success")
+    else:
+        flash('error!', category="error")
+    return jsonify({})
